Enable to delete task from the list of tasks, modularisation Fixes #16
